Shoot energy balls from player
It's not a bad system, can shoot every second. Animates well too!

We get this error when shooting:
Non-convex MeshCollider with non-kinematic Rigidbody is no longer supported since Unity 5.

The energy ball is a convex mesh collider though... So not sure here.
Tried to fix but it's taking up too much time